+++ src/main/java/net/frozenblock/wilderwild/mixin/block/leaves/LeavesBlockMixin.java
@@ -18,7 +18,11 @@
 
 package net.frozenblock.wilderwild.mixin.block.leaves;
 
+import com.llamalad7.mixinextras.injector.v2.WrapWithCondition;
+import com.llamalad7.mixinextras.sugar.Share;
+import com.llamalad7.mixinextras.sugar.ref.LocalBooleanRef;
 import net.frozenblock.wilderwild.block.impl.FallingLeafUtil;
+import net.frozenblock.wilderwild.tag.WWBlockTags;
 import net.minecraft.core.BlockPos;
 import net.minecraft.util.RandomSource;
 import net.minecraft.world.level.Level;
@@ -33,8 +37,26 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 public class LeavesBlockMixin {
 
 	@Inject(method = "animateTick", at = @At("HEAD"))
-	public void wilderWild$fallingLeafParticles(BlockState state, Level world, BlockPos pos, RandomSource random, CallbackInfo info) {
-		FallingLeafUtil.addFallingLeafParticles(state, world, pos, random);
+	public void wilderWild$fallingLeafParticles(
+		BlockState state, Level world, BlockPos pos, RandomSource random, CallbackInfo info,
+		@Share("wilderWild$usingCustomFallingLeaves") LocalBooleanRef usingCustomFallingLeaves
+	) {
+		boolean hasCustomParticles = FallingLeafUtil.addFallingLeafParticles(state, world, pos, random);
+		usingCustomFallingLeaves.set(hasCustomParticles);
+	}
+
+	@WrapWithCondition(
+		method = "animateTick",
+		at = @At(
+			value = "INVOKE",
+			target = "Lnet/minecraft/world/level/block/LeavesBlock;makeFallingLeavesParticles(Lnet/minecraft/world/level/Level;Lnet/minecraft/core/BlockPos;Lnet/minecraft/util/RandomSource;Lnet/minecraft/world/level/block/state/BlockState;Lnet/minecraft/core/BlockPos;)V"
+		)
+	)
+	public boolean wilderWild$fallingLeafParticles(
+		LeavesBlock instance, Level level, BlockPos blockPos, RandomSource randomSource, BlockState blockState, BlockPos blockPos2,
+		@Share("wilderWild$usingCustomFallingLeaves") LocalBooleanRef usingCustomFallingLeaves
+	) {
+		return !usingCustomFallingLeaves.get() || instance.builtInRegistryHolder().is(WWBlockTags.NON_OVERRIDEN_FALLING_LEAVES);
 	}
 
 }